Move `makedirs` to fs.py
