+++ dvc/utils/__init__.py
@@ -135,23 +135,6 @@ def copyfile(src, dest, no_progress_bar=False, name=None):
                     pbar.update(len(buf))
 
 
-def makedirs(path, exist_ok=False, mode=None):
-    path = fspath_py35(path)
-
-    if mode is None:
-        os.makedirs(path, exist_ok=exist_ok)
-        return
-
-    # utilize umask to set proper permissions since Python 3.7 the `mode`
-    # `makedirs` argument no longer affects the file permission bits of
-    # newly-created intermediate-level directories.
-    umask = os.umask(0o777 - mode)
-    try:
-        os.makedirs(path, exist_ok=exist_ok)
-    finally:
-        os.umask(umask)
-
-
 def _split(list_to_split, chunk_size):
     return [
         list_to_split[i : i + chunk_size]
