+++ tests/func/test_import.py
@@ -12,7 +12,7 @@ from dvc.exceptions import PathMissingError
 from dvc.exceptions import NoOutputInExternalRepoError
 from dvc.stage import Stage
 from dvc.system import System
-from dvc.utils import makedirs
+from dvc.utils.fs import makedirs
 from dvc.compat import fspath
 from tests.utils import trees_equal
